pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "premio-tronchi"
version = "0.1.0"
description = "Bot Telegram per la classifica dei tronchi al fantacalcio"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "python-telegram-bot==21.7",
  "pandas==2.2.3",
  "openpyxl==3.1.5",
  "requests==2.32.3",
  "beautifulsoup4==4.12.3",
  "python-dotenv==1.0.1",
]

[project.scripts]
premio-tronchi = "premio_tronchi.main:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
